feat: tagged-union models with discriminator-aware codecs
Replace the interface-based union pattern (with hand-written *_impl.go
helpers that don't survive regen) by a generated tagged-union struct
that handles JSON dispatch end-to-end.

For each oneOf schema with a discriminator (e.g. Peer, WireguardEndpoint,
NetworkGroupComponent), the generator now emits a struct holding the raw
JSON payload plus accessors:

  - Type() string                       — peeks "type" lazily via peekType
  - MarshalJSON / UnmarshalJSON         — round-trips the bytes verbatim
  - As<Member>() (<Member>, bool)       — checked accessor per variant
  - <Union>Variant interface            — for generic helpers
  - To<Union>() <Union> on each member  — wraps a concrete value

Concrete members (CleverPeer, ClientEndpoint, ...) get a custom
MarshalJSON that forces the discriminator field to its constant before
encoding, so json.Marshal(cp) is always a valid payload regardless of
whether the caller set cp.Type.

Why a struct instead of an interface:
  - []<Union> and *<Union> fields decode/encode without any custom
    UnmarshalJSON on the parent — the recursion footgun is gone.
  - Round-trip is byte-for-byte: we re-emit what we received, so unknown
    or future variants pass through cleanly.
  - Name collisions in the spec (e.g. NETWORK_GROUP vs NetworkGroup both
    PascalCase to "NetworkGroup") are handled by skipping accessor
    generation rather than emitting broken references to constants that
    may not survive the file overwrite race.

Adds models/json_helpers.go (peekType), generated alongside, kept in a
fixed file so it survives `make clean`. Adds round-trip tests
(union_unmarshal_test.go) and a runnable example
(peer_union_example_test.go) covering the full type-switch flow with
forward-compat for unknown variants.

Models regenerated from openapi.all.json — picks up two new schemas in
passing (versions_config, wannabe_otel_service).

+++ models/consumption_detail.go
@@ -4,6 +4,6 @@ package models
 
 // ConsumptionDetail
 type ConsumptionDetail struct {
-	Interval   string `json:"interval"`
-	Quantities []int  `json:"quantities,omitempty"`
+	Interval   string    `json:"interval"`
+	Quantities []float64 `json:"quantities,omitempty"`
 }
